// piClient/src/app/model/settings.ts
export class Settings {
  display: {
    weatherDefault: {
      show: boolean;
      index: number;
    };
    weatherHourly: {
      show: boolean;
      index: number;
    };
    weatherDaily: {
      show: boolean;
      index: number;
    };
    weatherLocal: {
      show: boolean;
      index: number;
    };
  };
}